fix free flight controls
When you're viewing through a camera rotated from some initial position to its
current position by yaw, pitch, and roll, and you want the camera to look
more to the right, it is not sufficient to say "add more yaw". This is because
what looks to be like more yaw from your perspective is some combination of
yaw, pitch and roll. This is why free flight wasn't working the way it should.
Using Rodrigues rotations we can rotate the camera relative to its current
orientation.